WIP test against a local vllm ollama server
sets up a vllm proxy to a locally running ollama server.
configures a connection and runs tests against it.

TODO: evaluation is not working yet.
The evaluation response is empty.

TODO: set a different eval model.

removes opentelemetry dumps to console
removes progress pushing state to the console both of which were polluting the logs