fix provider name
the provider name was incorrectly set to the model name after the config changes.
updates this to take the provider_name/model_name as the provider param for litellm.

fixes bug: when the llm response is `None`, a few attacks are failing due to unsafe
handling of `None` value.

removes excessive or repetitive logging